add browser module to i18next + localeSwitcher code refactoring
