+++ package.json
@@ -36,6 +36,7 @@
     "async-mutex": "^0.5.0",
     "clsx": "^2.1.1",
     "i18next": "^23.11.5",
+    "i18next-browser-languagedetector": "^8.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.51.4",
