+++ src/common/locales/by/translation.json
@@ -36,7 +36,7 @@
   "useTabsValuesParams": {
     "allDecks": "Усе",
     "myDecks": "Мае",
-    "favorites": "Улюбленыя"
+    "favorites": "Улюбёныя"
   },
   "paginationWithSelect": {
     "show": "Паказаць",
@@ -193,19 +193,19 @@
     "unknown": "Паспрабуйце зноў пазней"
   },
   "help": {
-    "help": "Падтрымка",
+    "help": "Дапамога",
     "step-go-home": "Націсніце тут, каб перайсці на галоўную старонку.",
     "step-select-language": "Націсніце тут, каб выбраць мову на старонцы.",
-    "step-change-theme": "Націсніце тут, каб змяніць тэму (светлая або тэмная).",
+    "step-change-theme": "Націсніце тут, каб змяніць тэму (светлая або цёмная)",
     "step-profile-drop-down": "Націсніце тут, каб адкрыць меню, у выпадаючаму меню вы можаце праглядзець свай прыватны профіль або выйсці з сваега акаўнта.",
     "step-search": "Вы можаце шукаць даркі па назвах.",
-    "step-tab-filter": "Фільтр для адлюстравання карчмі табліцы.",
-    "step-slider-filter": "Вы можаце адлюстраваць колоды ў табліцы па кількосці каратак. Каб зрабіць гэта, перамясцяйце гэтыя рулікі.",
+    "step-tab-filter": "Выберыце фільтр для адлюстравання калод у табліцы",
+    "step-slider-filter": "Вы можаце фільтраваць калоды ў табліцы па колькасці карт. Для гэтага перамяшчайце гэтыя паўзункі",
     "step-clear-filter": "Націсніце тут, каб ачысціць фільтры.",
     "step-add-new-deck": "Націсніце тут, каб дадаць новую дарку.",
     "step-this-is-table": "Гэта табліца, ў якой вы можаце глядзець свае карты, паўвайце увагу наверху ў загалоўках вы можаце глядзець назвы стулпцоў, пры націсканні на якіх вашы карты будуць адсартаваны.",
     "step-edit-table-deck-btns": "Тут вы можаце рэдагаваць сваю калоду (змяніць назву, зрабіць прыватнай, дадаць картку, выдаліць картку, вывучыць).",
-    "step-pagination": "Здзейсніце тут пераход па старонках і рэгульнаванне колькасці карт на старонцы.",
+    "step-pagination": "Тут вы можаце перамяшчацца паміж старонкамі і ўстанаўліваць колькасць карт на старонку",
     "step-private-deck": "Тут вы можаце знайсці значок маскі, які паказвае, што гэтая колода быў прыватнай і бачымай толькі вамі."
   },
   "controlButtons": {
@@ -214,6 +214,6 @@
     "last": "Апошні",
     "next": "Наступны",
     "open": "Адкрыць",
-    "skip": "Праскоч"
+    "skip": "Прапусціць"
   }
 }
