+++ src/components/Layout/Header/Header.tsx
@@ -7,9 +7,9 @@ import PersonOutline from '@/assets/icons/svg/PersonOutline'
 import CardMemoLogoGoldenPng from '@/assets/img/cardMemoLogoGolden.png'
 import DropdownMenuDemo from '@/components/ui/DropDown/DropDown'
 import DropDownItem from '@/components/ui/DropDown/DropDownItem'
+import LocaleSwitcherDrop from '@/components/ui/LocaleSwitcher/LocaleSwitcherDrop'
 import Typography from '@/components/ui/Typography/Typography'
 import ChangeTheme from '@/components/ui/changeTheme/ChangeTheme'
-import LocaleSwitcherDrop from '@/components/ui/localeSwitcher/localeSwitcherDrop'
 import { path } from '@/router/path'
 import { useLogoutMutation } from '@/services/auth/auth.service'
 import { MeResponse } from '@/services/auth/auth.types'
