+++ src/components/ui/LocaleSwitcher/i18nConfiguration.ts
@@ -1,6 +1,7 @@
 import { initReactI18next } from 'react-i18next'
 
 import i18n, { InitOptions } from 'i18next'
+import LanguageDetector from 'i18next-browser-languagedetector'
 
 import byTranslations from '../../../common/locales/by/translation.json'
 import enTranslations from '../../../common/locales/en/translation.json'
@@ -10,11 +11,21 @@ import uaTranslations from '../../../common/locales/ua/translation.json'
 
 const defaultLanguage = localStorage.getItem('locale') || 'en'
 
-const options: InitOptions = {
+const i18nextOptions: InitOptions = {
   debug: true,
   detection: {
-    cache: ['cookie'],
-    order: ['queryString', 'cookie'],
+    caches: ['localStorage', 'cookie'],
+    excludeCacheFor: ['cimode'],
+    order: [
+      'querystring',
+      'cookie',
+      'localStorage',
+      'sessionStorage',
+      'navigator',
+      'htmlTag',
+      'path',
+      'subdomain',
+    ],
   },
   fallbackLng: defaultLanguage,
   interpolation: {
@@ -39,5 +50,11 @@ const options: InitOptions = {
   } as const,
 }
 
-i18n.use(initReactI18next).init(options)
+i18n
+  .use(LanguageDetector)
+  .use(initReactI18next)
+  .init({
+    supportedLngs: ['by', 'en', 'kz', 'ru', 'ua'],
+    ...i18nextOptions,
+  })
 export default i18n
